Remove the old provider verification methods.
Now we have the claims package with the appropriate items, drop the old stuff.
Update the rest of the code to suit.
