+++ provider/provider.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"slices"
 	"sync"
 	"time"
 
@@ -30,6 +29,7 @@ type Provider struct {
 
 	cacheMu          sync.Mutex
 	cacheLastFetched time.Time
+	cachedJWKS       []byte
 	cachedHandle     *keyset.Handle
 
 	oidcDiscoveryURL string
@@ -37,7 +37,7 @@ type Provider struct {
 
 var _ jwt.Verifier = (*Provider)(nil)
 
-func (p *Provider) GetIssuer() string {
+func (p *Provider) Issuer() string {
 	return p.Metadata.issuer()
 }
 
@@ -49,97 +49,12 @@ func (p *Provider) Endpoint() oauth2.Endpoint {
 	}
 }
 
-func (p *Provider) JWKSHandle(ctx context.Context) (*keyset.Handle, error) {
+// JWKS returns the raw JWKS for this provider.
+func (p *Provider) JWKS(ctx context.Context) ([]byte, error) {
 	if err := p.refreshIfNeeded(ctx); err != nil {
 		return nil, err
 	}
-	return p.cachedHandle, nil
-}
-
-type IDTokenValidatorOpts struct {
-	ClientID       *string
-	IgnoreClientID bool
-	// ACRValues is the list of ACR values that are allowed for the token. The
-	// token will pass if its ACR claim is present and in the list. If empty, no
-	// ACR check is performed.
-	ACRValues []string
-}
-
-type IDTokenValidator struct {
-	tink *jwt.Validator
-	opts *IDTokenValidatorOpts
-}
-
-func (p *Provider) NewIDTokenValidator(opts *IDTokenValidatorOpts) (*IDTokenValidator, error) {
-	if opts == nil {
-		opts = &IDTokenValidatorOpts{}
-	}
-	vo := &jwt.ValidatorOpts{
-		ExpectedIssuer:   ptr(p.Metadata.issuer()),
-		ExpectedAudience: opts.ClientID,
-		IgnoreAudiences:  opts.IgnoreClientID,
-	}
-	tv, err := jwt.NewValidator(vo)
-	if err != nil {
-		return nil, fmt.Errorf("creating tink validator: %w", err)
-	}
-	return &IDTokenValidator{tink: tv, opts: opts}, nil
-}
-
-func (p *Provider) VerifyAndDecodeIDToken(ctx context.Context, token *oauth2.Token, validator *IDTokenValidator) (*jwt.VerifiedJWT, error) {
-	rawJWT, ok := token.Extra("id_token").(string)
-	if !ok {
-		return nil, fmt.Errorf("no id_token found in token")
-	}
-	j, err := p.verifyAndDecodeContext(ctx, rawJWT, validator.tink)
-	if err != nil {
-		return nil, err
-	}
-	if len(validator.opts.ACRValues) > 0 {
-		if !j.HasStringClaim("acr") {
-			return nil, fmt.Errorf("ACRs requested, but no ACR claim found")
-		}
-		acr, err := j.StringClaim("acr")
-		if err != nil {
-			return nil, fmt.Errorf("getting ACR claim: %w", err)
-		}
-		if !slices.Contains(validator.opts.ACRValues, acr) {
-			return nil, fmt.Errorf("jwt ACR %s not in requested list %v", acr, validator.opts.ACRValues)
-		}
-	}
-	return j, nil
-}
-
-type AccessTokenValidatorOpts struct {
-	ClientID       *string
-	IgnoreClientID bool
-}
-
-type AccessTokenValidator struct {
-	tink *jwt.Validator
-}
-
-func (p *Provider) NewAccessTokenValidator(opts *AccessTokenValidatorOpts) (*AccessTokenValidator, error) {
-	vo := &jwt.ValidatorOpts{
-		ExpectedIssuer:     ptr(p.Metadata.issuer()),
-		ExpectedTypeHeader: ptr(jwtTYPAccessToken),
-		ExpectedAudience:   opts.ClientID,
-		IgnoreAudiences:    opts.IgnoreClientID,
-	}
-	tv, err := jwt.NewValidator(vo)
-	if err != nil {
-		return nil, fmt.Errorf("creating tink validator: %w", err)
-	}
-	return &AccessTokenValidator{tink: tv}, nil
-}
-
-func (p *Provider) VerifyAndDecodeAccessToken(ctx context.Context, token *oauth2.Token, validator *AccessTokenValidator) (*jwt.VerifiedJWT, error) {
-	j, err := p.VerifyAndDecode(token.AccessToken, validator.tink)
-	if err != nil {
-		return nil, err
-	}
-	// TODO - add our additional checks here.
-	return j, nil
+	return p.cachedJWKS, nil
 }
 
 // Userinfo will use the token source to query the userinfo endpoint of the
@@ -179,21 +94,24 @@ func (p *Provider) Userinfo(ctx context.Context, tokenSource oauth2.TokenSource,
 }
 
 // VerifyAndDecode implements the tink [jwt.Verifier] interface. This provides
-// low-level verification of a JWT token against the provider's JWKS. Generally
-// the higher level [Provider.VerifyAndDecodeIDToken] and
-// [Provider.VerifyAndDecodeAccessToken] methods should be used. This will
-// always use a background context if network access to refresh the JWKS is
-// needed.
+// low-level verification of a JWT token against the provider's JWKS. For
+// verifying ID or Access tokens, the [lds.li/oauth2ext/claims] package should
+// be used.
+//
+// Note: This method may trigger network calls to refresh the JWKS if needed.
+// Using the [Provider.VerifyAndDecodeContext] method is recommended.
 func (p *Provider) VerifyAndDecode(compact string, validator *jwt.Validator) (*jwt.VerifiedJWT, error) {
-	return p.verifyAndDecodeContext(context.Background(), compact, validator)
+	return p.VerifyAndDecodeContext(context.Background(), compact, validator)
 }
 
-func (p *Provider) verifyAndDecodeContext(ctx context.Context, compact string, validator *jwt.Validator) (*jwt.VerifiedJWT, error) {
-	handle, err := p.JWKSHandle(ctx)
-	if err != nil {
+// VerifyAndDecodeContext provides low-level verification of a JWT token against
+// the provider's JWKS. For verifying ID or Access tokens, the
+// [lds.li/oauth2ext/claims] package should be used. This
+func (p *Provider) VerifyAndDecodeContext(ctx context.Context, compact string, validator *jwt.Validator) (*jwt.VerifiedJWT, error) {
+	if err := p.refreshIfNeeded(ctx); err != nil {
 		return nil, err
 	}
-	verif, err := jwt.NewVerifier(handle)
+	verif, err := jwt.NewVerifier(p.cachedHandle)
 	if err != nil {
 		return nil, fmt.Errorf("creating verifier: %w", err)
 	}
@@ -202,7 +120,7 @@ func (p *Provider) verifyAndDecodeContext(ctx context.Context, compact string, v
 		return nil, err
 	}
 	// perform an additional check, we should only ever pass tokens from the
-	// issuer.
+	// current issuer.
 	iss, err := jwt.Issuer()
 	if err != nil {
 		return nil, fmt.Errorf("getting issuer: %w", err)
